Industrial Foregoing Ore Laser Cleanup
Removed Sapphire and Ruby as they are not used.
Added Dimensional Shards (Rates TBD)
Unified all possible ores on EE ore chunks
Brought all remaining ores in as scripts for future weight/distribution changes if desired.
Added appropriate tags to Yellorium to bring it into the Ore Laser in case reactors are staying.
Added Oil as a fluid that can be mined with the new fluid laser.