NaturesAura Summoning Compat
ArsNouveau: Added Sylphs and Carbuncles. Both are used for automation aspects of ArsNouveau (carbuncles transport items similar to Rats, Sylphs slowly produce 'natural' items like logs and flowers). In order to obtain them, one must find them in the wild and perform a special action nearby. If more are desired, then more must be found. This ritual helps with the fact that they spawn very rarely, but does not diminish the need for the 'special action' once summoned in this way.

While adding these, I thought I'd toss in a few more for good measure.

Quark: Crabs, Frogs, and Foxhounds. These are also exceptionally rare in the world, so I thought it would be nice to have an alternate means of obtaining them. Once a few are obtained, they can of course be bred through normal means.

Thermal: I've added all of the elementals here. The recipe for spawning them is just as costly as summoning blaze through this ritual and realistically would not be very feasible for making a self sustaining summoner without a good looting sword.

A few more I considered, but figured I'd hold off on for now: Pillagers, Vindicators, Evokers, and Ravagers. Could be an interesting way of gaining access to them. Also the various new nether mobs (piglins, hoglins, piglin brutes). If you're OK with them, I can work out suitable recipes and add them in. At a very minimum, I'd suggest keeping the carbuncles and sylphs, however, since they're exceptionally rare in the world.