MythBoth generators - Disabled without breaking anything
Also reverted the pylons to default

I think the issue stems from how KubeJS is disabling the recipes. If I set a datapack to use more than one minecraft air as the ingredient, I get the same broken entries in patchouli, and I think kubejs is doing just that: replacing inputs with air.

So I've overwritten the recipe via the data pack method to disable it. There's also an extra note in patchouli stating it's been disabled intentionally.