quest-loot-tables
These are the quest loot tables for Miner's Delight and Sorcerer's Delight that will replace the FTB quest loot boxes in place. Primarily this is to allow random enchanted items in these General Quest rewards.

To call them the quest reward needs to be set to Command and the following inserted:

/execute at @p run loot spawn ~ ~1 ~ loot enigmatica:chests/quest_sorcerers_delight

/execute at @p run loot spawn ~ ~1 ~ loot enigmatica:chests/quest_miners_delight

This will drop the items on the player, as discussed, using the loot/give command instead can potentially lead to items being lost.