more sawing updates
Alright, moved the variable into global constants. I'd like to move the switch case logic there too at some point and just have the constant consist of an array of arrays with all the proper 'types' in it for each wood. That's going to take me more time though and for now I'd rather get this updated for tonight.

This also includes some missing wood in Thermal and Pedestals.