Loot table remake
Loot tables for all current quests.

For future reference, the following two sites were of enormous help:

https://amaury.carrade.eu/minecraft/loot_tables
This allows for building the loot tables right in the page and intelligently escapes double quotes in NBT strings

https://gaming.stackexchange.com/questions/334946/minecraft-1-13-displayname-nbt-tag
For all of the named Tetra items, the first site was having trouble. I had to wrap the diplay name in single quotes like so:

from this: `display:{Name:\"{\"text\":\"Blackbog's Sharp\"}\"}`

to this: `display:{Name:'{\"text\":\"Blackbog’s Sharp\"}'}`

And that also required using a curly apostrophe for possessives as a plain single quote was breaking the outside wrapping.