Add custom values to names.cfg
Currently the Apotheosis bosses all have default equipment and names, excluding netherite which has been removed from being possible.

This proposed change adds a number of materials to the pool by giving them names, trying not to be overpowered and appreciating the current values for what non-bosses may spawn with. This also adds a bunch of prefixes to boss names to add a bunch of variety (I'm tired of every boss I meet having the prefix Doctor, for example) and new name-parts for some additional random names.

As a side effect, I believe this will allow some of the named materials to appear as affix dungeon loot, but this is in my opinion desirable - as long as the materials are not overpowered compared to existing affix dungeon loot.