Occultism Ore Generator Revamp
Complete overhaul of ores that are generated to remove unused ores and to make it generate chunks. Weights are in line with the rest of our ore generators.

Added Iesnium to make the ritual somewhat renewable, as otherwise the only way to continue once it wears out is to manually mine more.

Increased durability on the 'advanced' version, slightly sped up, and increased yield per operation. Sadly, none of this works yet: https://github.com/klikli-dev/occultism/issues/79