Change Cobbleworks so it doens't interfere with basic cobble generators.
Returns cobble/stone/basalt generation back to generic vanilla
re-introduces cobbleworks tags on new catalyst blocks. Iron Block for cobble, and diamond block for stone.

removed blackstone from the generic cobbleworks as it is easily found in the nether and is used for progression. #833

Added Purpur Stone and Myalite to both cobblegens.