Additions cleanup
First step in cleaning this up. Many of our functions are re-calling the same lookups (preferred item of ingot, ore, dust, etc), so I've moved that out of the functions and into the main material loop. This way each item type is checked once and passed in through the function as needed.

Phase two would be to clean up the scripts that have better KJS integration now, like Thermal, IE, Mekanism, and Create, just as the previous commit reworked the Blood Magic scripts.