Additions cleanup - phase 2
re-factor of additions.js to use new kjs syntax for custom recipes and addon recipe handlers where available.

Generally tried to standardize the scripts to a similar layout for readability.

This also removes IE hammer>plate recipes as EE now covers that. Next step will be to rmove the IE hammer > dust recipes and standardize the EE recipes on tags.