changed loot table for wild hunt ritual
added a bit to the loot tables for this ritual as it is fairly expensive(gold block, silver block and copper block + a diamond and cant be fully automated) and the only perk to it before was guaranteed wither skulls which are pretty easy to get in this pack. There is a high chance of nether fortress loot(nothing special at all), a decent chance at a random enchanted book enchanted at lvl 20-30, a pretty rare chance at an apo tome at lvl 50, and a suuuuuuuuper rare chance at a random artifact(right around 5%), along with the guaranteed skull.