Add Create Schematics
This is the first time I've ever worked with Github beyond submitting issues, so please bear with me.

This is a batch of Create Schematics, mirroring the existing building_gadgets_patterns folder. I've included a README.txt to help explain how the schematics work; none of them will show up in-game by default, but if they are copied back to the root schematics folder (clientside), they will be available in-game at the schematic table. Nice to have an alternate method to the Building Gadget, per a discussion I had with Motoko on Discord.