Market Update and Chorus Insolator Fix
Chorus wasn't renewable in the Insolator since it didn't output Chorus Flowers. Fixed.

Adds the Better End Forge seeds and saplings to the Market, though at a cost of Amber instead of Emeralds to keep them tied to the End.

Adds missing Upgrade Aquatic River Sapling to the market as well.