Add Recipes for Artifact Level Infinity Tools
I think there's a couple ways to go here, but I'm starting with this.

Running these recipes pretty much requires a high heat fusion reactor to do in a timely manner. Not to mention actually producing the Antimatter required (100k Uranium Ingots worth).

All in all, the craft should take about 20 minutes if the power costs can be kept up.

I also opted to actually use any tier of Infinity tool in the craft. I was originally leaning towards converting more conventional items, like a vanilla Trident for the Infinity Trident. A Saw blade for the saw, drill head for the drill, Netherite hammer for the hammer, and an Ender Chest for the backpack... Something similar may still be desirable since this -will- result in lost items if someone were to put their old Infinity backpack into the recipe.

#1694