PVJ, Farmers Delight, and BYG Integration Update
#1840
Creates recipes for PVJ ground clutter items and adds crop recipes for the plants.

Adds a note to red and brown mushrooms and colonies that explains how to obtain the colonies.

Added new crop recipes for the Imparius mushroom trees.
Added Imparius Phylium as valid soil for botany pots.

Tag all the things.