Make Botany Pots and Terracotta washable and Botany Pots re-dyeable and other improvements
Implements recipes to dye any botany pot to any other color, as well as washing recipes to revert to the plain terracotta look.

#1875

Also expands existing Mek Injecting recipe to work with any color terracotta (converts to clay) and adds a TPP recipe to do the same.

Reduces processing time of Essence > Memory Essence in TPP. Adds reverse recipe. Same for Dissolution chamber.

As a note, the washables tag was implemented in order to prevent recipes such as "Plain Terracotta washing to Plain Terracotta". In particular, for Create's washing this seems like something that would be bad for TPS since it has to keep checking the Plain Terracotta and attempting to re-wash it so long as it remains in the wash stream.