Tetra Scriptification
Moves Tetra materials and replacements into scripts where they can be more easily manipulated.

This also brings in all base Tetra materials, allowing them to be modified more simply as well. Left some notes for future material improvements in comments.

As proof of concept, all materials have been granted a 20% bonus on durability and magic capacity. All seems to be working well.

For simpler management in the future, we may want to consider setting up set 'tiers' for each stat. This way we could say, for instance, that Iron is Tier 3 durability and Tier 2 Speed, while Gold is Tier 1 durability and Tier 3 speed, as a rough example. This way we can have a central place to re-balance all tools against each other.

I've opted to delete the JSONs in this PR even though they'll likely remain on a player's instance. The JSON injection is over-writing those same files, however, so it  makes no difference. I was seeing the changes from the scripts even with them in place. So this is just a bit of cleanup.