Botania Orechid Rebalance Attempt
Having a bit more trouble with this than expected. It's getting late so I think I'm maybe just looking at it cross-eyed.

The scripts do not appear to overwrite the default values, however. I've tried purely with the script as written (with the key component uncommented, of course) and I've tried a straight datapack injection in the data folder. Neither worked :(

In both cases, I'm still getting ruby/sapphire and no fluorite is being generated.

Care taking a look to see if I've missed something?