DungeonCrawl Winner Dungeons: Initial Loot Table Creation
I wanted to take a little different approach to this, given the size of these tables, so I've put it into a script that makes it easier to read/edit.

These tables are not yet ready to be implemented, but at least get the tables in and ready for testing with the actual dungeons.

Proof of concept is in dungeon_ocean_common that is mostly configured and has dynamically injected functions. dynamic NBT is working as well.