Remove the need for the Ars Nouveau glyph recipe, use a not filter instead.
The Ars Nouveau glyph recipe was added because the nether brick consolidation replaced nether bricks with a tag, which the glyph recipes do not support. This is now superfluous as KubeJS now supports a "not" filter on recipe types, meaning that the recipe type can simply be excluded from consolidation instead.