Update unify_materials.js
Proof of Concept Magical 5x Ore Processing.
All art assets are temporary. To be replaced later.

Drop the ore into a Mana Pool (requires Creational Catalyst under it)
![image](https://user-images.githubusercontent.com/9543430/119401142-1acbf080-bca9-11eb-95a2-54828085a240.png)

Zap result with lightning and a Rune of Asgard catalyst
![image](https://user-images.githubusercontent.com/9543430/119402991-b52d3380-bcab-11eb-940a-8f6c137617be.png)
Return chance doesn't appear to work with lightning. On the other hand, it does let you craft many sets with only 1 run as input. So it will be up to the player to decide how efficient they want to be. Drop 1 rune and 64 inputs? more? Considering the recipe for the Ingot of the sky used in this, it'll be unwise to do small batches.

This step produce roughly 3x output with a mix of the input item and a secondary such as Nickel from Iron.

Crumble in Natural Infusion Altar
![image](https://user-images.githubusercontent.com/9543430/119402151-86628d80-bcaa-11eb-9629-1f56f5753507.png)

Wash in Starlight:
![image](https://user-images.githubusercontent.com/9543430/119402204-9d08e480-bcaa-11eb-9d1d-e8564e8088af.png)
Chance to use up both starlight and the winter rune.

This is the final multiplication step, producing roughly 15x of the input material. 9 of the output goes into making an ingot.

Finally, combine in a Runic Altar. This rune, as with all runes used in Altar crafting, is always returned. Calx of the End is something that will likely be on auto craft already given it's many uses.
![image](https://user-images.githubusercontent.com/9543430/119403137-e279e180-bcab-11eb-92ea-433f53608d6d.png)