Latex Unification
IF Latex is now the only latex available.
The Arboreal Extractor can extract it endlessly from certain living trees or the IF Fluid Extractor can be used to pull it out of cut logs as normal.

'Easy' recipes using a bucket and vines/dandelions now produce 3 tiny dry rubber.
Latex can still be pressed out of vines/dandelions with a multi-press at a rate of 50mb. So 18 of these items makes 1 dry rubber, as opposed to 24 with the bucket method.
Fractionating Still produces 1 Dry Rubber  from 900 mb of latex. In line with the IF Latex Processing Unit

Completes #1592