Add Hive_Upgrade_Recipes
Moving Hive & Apiary recipes to the custom hive_upgrade_recipe handler for Resourceful Bees. This still registers a crafting table recipe, but if there are bees in the input hives it will return them in a bee box.

I did discover an (unintentional) caveat to this; currently ALL apiaries/hives require NBT from placement for this to work properly. I intend to add JEI descriptions/tooltips as a warning.

This sounds like a bug that I THINK the RB team is going to fix. I can inject NBT into the outputs here to work around this, but I'm a little concerned about long-term complications for autocrafting; I've left a commented-out example in the first recipe in the "normal" file for future reference.