Implement Hide Recipe from JEI list
Thanks @MaxNeedsSnacks! Works beautifully.

This creates a list for hiding the placeholder recipes I've been using to prevent Patchouli breakages in JEI. The recipes still show properly in their respective books with the placeholder items directing the player to look in JEI (or instructing them that the recipe is simply disabled), but they are no longer visible through JEI, removing that clutter. Huzzah!