Add Molten/Liquid Bees
Crafting with Tinkers will require fluids for the bees; this creates a fluid registry in startup_scripts, and adds the fluids.

Based on the example code on Lat's wiki, I SHOULD be able to use .displayName() rather than en_us.json to add the names, but this isn't working so I've commented it out and used the lang file instead. I'm trying to work with Max to get this working, but it works fine as is.

I've also left in a third (commented out) check for fluids using custom textures. Not sure if we'll need it, but it's there if we do.