Alternate PNC Recipes: Elevator Frames and Canon Barrels
Was brought up in Suggestions that these elevators are very iron heavy to craft, and they kinda are. Especially the frames required to fill all spaces above the elevator bases in the shaft.

So this introduces two cheaper recipes to assist as a late game bonus on the assembly table.

Elevator Frames are normally 6 compressed iron for 4. The assembly table makes them 4 compressed iron for 8.

Cannon Barrels used in the cylinders are normally 6 reinforced brick walls and a pressure tube for 1. This give 2 barrels for 6 walls.