Fix for some workbenches not being seen as workbenches.
Restandardize on `forge:workbenches` rather than current `forge:workbench` as tcon appears to be setting that standard now.

Fix MCTB workbenches not being in workbenches tag

Resolves #2763

Deleting old Workbench JSON as it's moved into Shaped recipes now with an ID override. Guess I missed this one in the un-json-ifying. Should have no impact on players as the script will override it if present.