Industrious & Thermal Bees (and other updates)
Adding the Industrious Bee - crafted in the IF dissolution chamber (to a bee in a jar). Combs produce liquid latex, pink slime balls, and Meaty Honey. Meaty honey adds lots of saturation with a chance of a few other effects. The bee has a number of mutations that feel on-theme for IF.

Add Blizz/Blitz/Basalz bees. They still need textures, but they have the appropriate particle effects. They are all mutated from the proper mob by the Blaze Bee, and their combs put out rods and appropriate material chunks or snowballs.

Updated the Wasabee - it can now be created by mutating a Cat with a Water Bee - finding a Kitten Bee to breed one is a pretty hard gate. This does mean it is possible a mutation will occur in-world, if a water bee happens to fly over a village that has a cat.

Updated Resourceful Bees constants to add a bunch of the things I've added to RB but not added to their related constants.