Create Sequenced Assembly Script
Fixes #2835.

The default Create recipe for the Precision Mechanism has a chance output for "create:golden_sheet", which seems to be invalidating the recipe since we've removed it.

Implemented a clean-ish script for doing Sequenced Assembly, and am overriding all 3 default recipes to fix the above issue and remove the chance of "salvage" - for these at least the idea of chance failure seems like a super frustrating mechanic for a factory.