[Expert] Armor crafting overhaul
This moves armor out of basic crafting table recipes and into the realm of tech and magic. The goal here is to make certain sets a bit tougher to craft, for instance diamond and netherite, making found armor from exploration a bit more valuable.

Most armors follow a simple assembly crafting recipe such as this:

![image](https://user-images.githubusercontent.com/9543430/131402263-7718a999-a82d-480b-b231-ce3e11d58e90.png)

Note that this is _not_ tiered crafting in the sense that you'll need to craft through every armor to get the high value stuff. Instead this just adds a few layers to most armor sets. Plate style armor, such as Iron and other metals is typically crafted on a base of Chain which is built by hand in a crafting grid still:
![image](https://user-images.githubusercontent.com/9543430/131402494-31d12194-e8be-45bd-83bf-3e899de40775.png)

Some, such as the NA armors, use leather instead as the base, while robe type armors weave threads into other woven armors:
![image](https://user-images.githubusercontent.com/9543430/131402614-d7636b91-9b1f-4323-8b64-546900d92a8c.png)

![image](https://user-images.githubusercontent.com/9543430/131402628-df020b62-01c2-435c-b9ce-c4c5313bb574.png)

![image](https://user-images.githubusercontent.com/9543430/131402640-962da11d-52b6-468e-8ac9-f41a11f5f3e5.png)

Others still, such as pneumatic armor, become a more involved process, requiring both the Create assembly system and the PNC assembly table to complete.

![image](https://user-images.githubusercontent.com/9543430/131402832-ee2d4a65-c339-41ba-b538-8153b94f48a3.png)

![image](https://user-images.githubusercontent.com/9543430/131402888-d7fb2f05-574d-4e33-b37f-0906434ab673.png)

![image](https://user-images.githubusercontent.com/9543430/131402863-c390ce04-c49a-4ea2-9d8f-301b2aaa0e18.png)