[Expert] Godforged tweaks
Alters Godforged Blocks to look for a tag, rather than a specific item when breaking them for relics.

For normal mode, this is only applied to the Nebu Hammer, the normal tool for this. For Expert, however, I've added the Alf Shatterer and Alfsteel AIOT to the list.

This just opens up a few extra possibilities for automation. May back out the previous Pharaoh summon. Thoughts?

Also updates NA quest chapter to cover some extra details concerning the Aura Field Creator.