[Expert] Concrete Cooling
Concrete Cooling
![image](https://user-images.githubusercontent.com/9543430/132759753-3263aaa4-a3bc-4d58-8900-869f385d76f5.png)

Bit of a bonus for keeping it cooler. Get it down to 20c for a 10% increase in yield.
![image](https://user-images.githubusercontent.com/9543430/132759848-044c613b-0a65-4f07-ae97-952edead12e6.png)

That's not particularly cold, but concrete doesn't usually like cold weather curing either.

And JEI update for concrete cooling in world:
![image](https://user-images.githubusercontent.com/9543430/132760358-048adbbd-77cf-4889-9129-bbbcd02d19fb.png)
The default IE cooling happens faster than this though, so just for the sake of consistency.