Refactor Ponder base.js
Before things get out of control, I'm refactoring how bits of the Ponder structure are arranged. Ponder's base.js now has a constant for the various categories, with the items that should be displayed associated; the scripts iterates through this to create the categories, and then iterates through it to register the items to the categories. This removes all of that code from the various Ponder scripts.