Recategorize Things
After discussing it in the Ponders thread, I think we've settled on 5 categories - Portals, "Basics", Tech, Tools, and Magic, to mirror (somewhat) the way quests are laid out. We don't have any magic ponders yet, so no category for it yet. I've rearranged the layout for the scripts and base.js to mirror.