Hive Upgrade Type/Tier Recipes
Moves to using type/tier ingredients added for RBees 0.9.9.6 in recipes; this will allow hives upgraded in-world to be used as their tier for crafting recipes. The only downside is that this can confuse autocrafting a bit - technically, for example, a T3 Beehive that is actually T2 is a thing (although unobtainable), which makes it possible for RS to fail due to recursive recipes. This seems to only happen if it doesn't have recipes for the whole crafting chain, however.