[Expert] Mineral Survey Drones
Requires pneumaticcraft-repressurized-1.16.5-2.14.3-255 or later. No official release yet, so may need to wait.

Introduces per dimension trades. As a starting point, I wanted to set up a way to get some dimension specific resources from Atum and Undergarden. I'm not 100% sure of the actual loot table, might be better to limit potential items a bit more... Perhaps to only the interesting materials from the dimension, or giving those a much higher weight. Thoughts?

As an initial balance, these trades are available 5 times per day and can only be requested from the appropriate dimension. Each one requires a steel paxel, filled air canisters, and a few other disposable items, as well as a chest everything will go in.

You have a 15% chance of getting a very damaged paxel back and 100% chance of getting the emptied air canisters back.

The chest then pulls a loot table with a selection of ores at different weights, which appear in the chest when placed.

![image](https://user-images.githubusercontent.com/9543430/134227821-414aedbb-6120-4be4-9722-38674ec8b20b.png)

https://streamable.com/isisiu