Update amadron_survey_loot_tables.js
Updated weights
Worked them out so the the percentage drop can be seen more clearly from the weight. So 50% of the materials in either table will now be the interesting stuff from those dimensions and the weights all add up to 10k so it's easier to see what percent each item is. 17% for nebu, for instance.