[Expert] Swap Tremblecrust for Shiverstone
Minor change, but technically tremblecrust isn't meant to be obtainable in survival. The only reason we can is because it's part of the random stone generator.

So, to avoid confusion, this swaps out tremblecrust for shiverstone instead. This can be found in small pockets in the undergarden like andesite in vanilla.