[Expert] PNC Fixes
Seems there's a bug with PNC's refinery that makes it fail to recognize a recipe if it initially receives less than is required to start. Example, lithium needs 100mb of brine to start, but the water > brine recipe outputs 10.

This leads to there being 10 brine in the second refinery and as it fills to 100+, it never rechecks the recipe.

Fix is to simply match the speeds so the first outputs 10 and the second needs 10 to start.

The recipe handler also requires 2 or more fluid outputs. I was cheesing this before by outputting 9/1 in two outputs. But with the lowered input for brine to lithium, I cannot do this. So now both machines output waste steam.

Renamed Liquid Steam to Dense Steam.