Force IPv4 ping in start-automated-server.bat
Ping does not respond with a TTL if the server responds with an IPv6
address. This essentially means that the script will think the user does
not have internet, if they have a public IPv6 address. Forcing IPv4 in
the ping command, ensures TTL gets returned, and the startup can
proceed.

Closes #4043