[Expert] Bring Dandelifeon earlier
Basically trading this for the Rosa Arcana, which is the intended flower for late game mana madness. This one generates quite a bit, but at a very low rate in terms of mana per tick. Seems fair to bring it forward considering the room it takes to utilize it and the increased mana use throughout mid game.