[Expert] Re-balance aura recipes
The natural altar has a cap of 300 aura per tick that it can absorb from the area. With this in mind, a lot of our recipes would end up fully draining the altar, then resulting in a drastic slowdown as it couldn't draw in enough to meet the intended rate.

This re-balances those values a bit to help with this. It also introduces a new recipe to make Spirits of Birthing to resolve #4806

Also adds a tooltip to make it clear that they can still be obtained naturally through manual breeding. If people figure out the trick to automating breeding, so much the better. But this means there's a way forward for folks that miss it.