Remove Mushroom Colonies from recipes
These are not fun to automate by any means and would require the player to do some really off the wall things to make in the quantities needed for the Mastery Shards.

Also reduces the time it takes to craft a few other cooking pot recipes. Notably Ender Calx and Candles. Candles now need string as the secondary item rather than in the pot itself.