Add direct craft for Forgotten tools
Currently the only way to obtain them is via the smithing table, which is fine if you just want one... but end game requires making many of them for irradiated ore miners for the excavation mastery shard.

It's a bit more expensive than doing it by hand, but this gives the players an automation friendly path to making these tools.